refactor(literature): split controller + export into focused modules; review nits
Address review feedback on #824. Behavior-neutral: each split file becomes a
re-export barrel so no call site (API route or check harnesses) changes.

Structure:
- literatureSearch.controller.ts (2163L) -> barrel + counting / records / llm
- literatureExport.ts (407L) -> barrel + blocks / strategyDoc / synthesisDoc / sheets
- db/wcmExperts: SQL moved to wcmExperts.repository.ts; controller now a thin orchestrator

Nits:
- literatureAllowlist: parse once into a Set (O(1)); trim the incoming cwid so a
  stray-whitespace CWID no longer silently misses
- literatureDocx: render -> renderBlock, RenderNode type, exhaustive never-default,
  hoisted SPACING / DEFAULT_STYLES consts
- check/question/recall/screen harnesses: guard the .env.local read with existsSync;
  wrap the compile execSync in a descriptive error

tsc --noEmit clean across the project; check:literature:pure green.
